Type cast

+++ trunk/src/grdfilter_func.c
@@ -1,5 +1,5 @@
 /*--------------------------------------------------------------------
- *	$Id: grdfilter_func.c,v 1.25 2011-06-02 18:57:50 guru Exp $
+ *	$Id: grdfilter_func.c,v 1.26 2011-06-05 17:23:27 jluis Exp $
  *
  *	Copyright (c) 1991-2011 by P. Wessel, W. H. F. Smith, R. Scharroo, and J. Luis
  *	See LICENSE.TXT file for copying and redistribution conditions.
@@ -278,7 +278,7 @@ GMT_LONG init_area_weights (struct GMT_CTRL *GMT, struct GMT_GRID *G, GMT_LONG m
 
 		GMT_col_loop (GMT, A, row, col, ij) {
 			col_weight = dx * ((A->header->registration == GMT_GRIDLINE_REG && (col == 0 || col == (A->header->nx-1))) ? 0.5  : 1.0);
-			A->data[ij] = row_weight * col_weight;
+			A->data[ij] = (float)(row_weight * col_weight);
 		}
 	}
 #ifdef DEBUG
@@ -846,7 +846,7 @@ GMT_LONG GMT_grdfilter (struct GMTAPI_CTRL *API, struct GMT_OPTION *options)
 					if (slow) {	/* Add it to the relevant temporary work array */
 						if (slower) {	/* Need to store both value and weight */
 							work_data[n_in_median].x[0] = Gin->data[ij_in];
-							work_data[n_in_median++].x[1] = weight[ij_wt] * A->data[ij_in];
+							work_data[n_in_median++].x[1] = (float)(weight[ij_wt] * A->data[ij_in]);
 						}
 						else	/* Only need to store values */
 							work_array[n_in_median++] = Gin->data[ij_in];
@@ -859,9 +859,9 @@ GMT_LONG GMT_grdfilter (struct GMTAPI_CTRL *API, struct GMT_OPTION *options)
 						n_conv++;	/* Points used inside filter circle */
 						if (Ctrl->A.active) {	/* Store selected debug info in Gin data */
 							if (Ctrl->A.mode == 'a') Gin->data[ij_in] = A->data[ij_in];
-							else if (Ctrl->A.mode == 'w') Gin->data[ij_in] = weight[ij_wt];
-							else if (Ctrl->A.mode == 'r') Gin->data[ij_in] = weight[ij_wt];
-							else if (Ctrl->A.mode == 'c') Gin->data[ij_in] = w;
+							else if (Ctrl->A.mode == 'w') Gin->data[ij_in] = (float)weight[ij_wt];
+							else if (Ctrl->A.mode == 'r') Gin->data[ij_in] = (float)weight[ij_wt];
+							else if (Ctrl->A.mode == 'c') Gin->data[ij_in] = (float)w;
 						}
 #endif
 					}
